Fix to prevent cross-contamination between REst & XML.
The setClient method will be called on the Rest Connector rather than the XML connector if rest
credentials are present. For other methods the choice of connector also includes the presence of the function,
since xml has some functions & rest others this kinda works. However, for setClient it exists on both
so we need to specify