Adjust GroupMember code for updated phpleague/csv plugin
The update to csv 9 is pretty breaking!

Here are the docs
https://csv.thephpleague.com/9.0/upgrading/

Relevant here are
- I switched to using a statment which incorporates the LIMIT param at the point of definition
(rather than when iterating). I expect that not using limit would have memory implications.
- The ability to pass a formatting  callable is gone. Stream processors 'seem' to be the
replacement but it started to look like a lot of  hard work without being sure it WAS the right
approach so instead I moved the field wrangling that it was doing to the Contact object
and  the casting to the contact object to CRM_Omnimail_Omnigroupmembers

The only other place instantiating the Contact object is the PrivacyInformationRequest class
which only passes in the opt_in_timestamp - so this should be pretty safe